Refactor engines into modular utility packages
